Refactor AgentChatMessages modules
- extract AgentChatMessages helpers, types, and scroll logic into submodules
- keep the scroll behavior and public exports intact
- allow teamwork scaffold tests to opt into git worktrees
